Rename files and delete old lexer
