--- lexer/test1.c
@@ -1,8 +1,0 @@
-/* comment */
-int j = 3;
-int main () {
-  int i = j;
-  int k = 3;
-  int p = 2 * j;
-  assert p == 2 * i;
-}
