--- lexer/test1_result
@@ -1,35 +1,0 @@
-LexToken(INT,'int',2,14)
-LexToken(ID,'j',2,18)
-LexToken(EQUALS,'=',2,20)
-LexToken(INT_CONST,'3',2,22)
-LexToken(SEMI,';',2,23)
-LexToken(INT,'int',3,25)
-LexToken(ID,'main',3,29)
-LexToken(LPAREN,'(',3,34)
-LexToken(RPAREN,')',3,35)
-LexToken(LBRACE,'{',3,37)
-LexToken(INT,'int',4,41)
-LexToken(ID,'i',4,45)
-LexToken(EQUALS,'=',4,47)
-LexToken(ID,'j',4,49)
-LexToken(SEMI,';',4,50)
-LexToken(INT,'int',5,54)
-LexToken(ID,'k',5,58)
-LexToken(EQUALS,'=',5,60)
-LexToken(INT_CONST,'3',5,62)
-LexToken(SEMI,';',5,63)
-LexToken(INT,'int',6,67)
-LexToken(ID,'p',6,71)
-LexToken(EQUALS,'=',6,73)
-LexToken(INT_CONST,'2',6,75)
-LexToken(TIMES,'*',6,77)
-LexToken(ID,'j',6,79)
-LexToken(SEMI,';',6,80)
-LexToken(ASSERT,'assert',7,84)
-LexToken(ID,'p',7,91)
-LexToken(EQ,'==',7,93)
-LexToken(INT_CONST,'2',7,96)
-LexToken(TIMES,'*',7,98)
-LexToken(ID,'i',7,100)
-LexToken(SEMI,';',7,101)
-LexToken(RBRACE,'}',8,103)
