--- lexer/test2.c
@@ -1,2 +1,0 @@
-for (int i = 0; i < 100; i++)
-    (*a)[i];
