--- lexer/test2_result
@@ -1,22 +1,0 @@
-LexToken(FOR,'for',1,0)
-LexToken(LPAREN,'(',1,4)
-LexToken(INT,'int',1,5)
-LexToken(ID,'i',1,9)
-LexToken(EQUALS,'=',1,11)
-LexToken(INT_CONST,'0',1,13)
-LexToken(SEMI,';',1,14)
-LexToken(ID,'i',1,16)
-LexToken(LT,'<',1,18)
-LexToken(INT_CONST,'100',1,20)
-LexToken(SEMI,';',1,23)
-LexToken(ID,'i',1,25)
-LexToken(PLUSPLUS,'++',1,26)
-LexToken(RPAREN,')',1,28)
-LexToken(LPAREN,'(',2,34)
-LexToken(TIMES,'*',2,35)
-LexToken(ID,'a',2,36)
-LexToken(RPAREN,')',2,37)
-LexToken(LBRACKET,'[',2,38)
-LexToken(ID,'i',2,39)
-LexToken(RBRACKET,']',2,40)
-LexToken(SEMI,';',2,41)
