--- lexer/test3.c
@@ -1,4 +1,0 @@
-int a=10;
-int *p;
-p = &a;
-int b = *p;
