--- lexer/test3_result
@@ -1,20 +1,0 @@
-LexToken(INT,'int',1,0)
-LexToken(ID,'a',1,4)
-LexToken(EQUALS,'=',1,5)
-LexToken(INT_CONST,'10',1,6)
-LexToken(SEMI,';',1,8)
-LexToken(INT,'int',2,10)
-LexToken(TIMES,'*',2,14)
-LexToken(ID,'p',2,15)
-LexToken(SEMI,';',2,16)
-LexToken(ID,'p',3,18)
-LexToken(EQUALS,'=',3,20)
-LexToken(ADDRESS,'&',3,22)
-LexToken(ID,'a',3,23)
-LexToken(SEMI,';',3,24)
-LexToken(INT,'int',4,26)
-LexToken(ID,'b',4,30)
-LexToken(EQUALS,'=',4,32)
-LexToken(TIMES,'*',4,34)
-LexToken(ID,'p',4,35)
-LexToken(SEMI,';',4,36)
