--- lexer/test4_result
@@ -1,26 +1,0 @@
-LexToken(INT,'int',1,0)
-LexToken(ID,'v',1,4)
-LexToken(LBRACKET,'[',1,5)
-LexToken(INT_CONST,'5',1,6)
-LexToken(RBRACKET,']',1,7)
-LexToken(EQUALS,'=',1,9)
-LexToken(LBRACE,'{',1,11)
-LexToken(INT_CONST,'1',1,13)
-LexToken(COMMA,',',1,14)
-LexToken(INT_CONST,'3',1,16)
-LexToken(COMMA,',',1,17)
-LexToken(INT_CONST,'5',1,19)
-LexToken(COMMA,',',1,20)
-LexToken(INT_CONST,'7',1,22)
-LexToken(COMMA,',',1,23)
-LexToken(INT_CONST,'9',1,25)
-LexToken(RBRACE,'}',1,26)
-LexToken(SEMI,';',1,27)
-LexToken(ASSERT,'assert',2,29)
-LexToken(ID,'v',2,36)
-LexToken(LBRACKET,'[',2,37)
-LexToken(INT_CONST,'3',2,38)
-LexToken(RBRACKET,']',2,39)
-LexToken(EQ,'==',2,41)
-LexToken(INT_CONST,'7',2,44)
-LexToken(SEMI,';',2,45)
