--- lexer/test5.c
@@ -1,1 +1,0 @@
-float (*foo)[3] = &(float[]){ 0.5, 1., -0.5 };
