--- lexer/test5_result
@@ -1,24 +1,0 @@
-LexToken(FLOAT,'float',1,0)
-LexToken(LPAREN,'(',1,6)
-LexToken(TIMES,'*',1,7)
-LexToken(ID,'foo',1,8)
-LexToken(RPAREN,')',1,11)
-LexToken(LBRACKET,'[',1,12)
-LexToken(INT_CONST,'3',1,13)
-LexToken(RBRACKET,']',1,14)
-LexToken(EQUALS,'=',1,16)
-LexToken(ADDRESS,'&',1,18)
-LexToken(LPAREN,'(',1,19)
-LexToken(FLOAT,'float',1,20)
-LexToken(LBRACKET,'[',1,25)
-LexToken(RBRACKET,']',1,26)
-LexToken(RPAREN,')',1,27)
-LexToken(LBRACE,'{',1,28)
-LexToken(FLOAT_CONST,'0.5',1,30)
-LexToken(COMMA,',',1,33)
-LexToken(FLOAT_CONST,'1.',1,35)
-LexToken(COMMA,',',1,37)
-LexToken(MINUS,'-',1,39)
-LexToken(FLOAT_CONST,'0.5',1,40)
-LexToken(RBRACE,'}',1,44)
-LexToken(SEMI,';',1,45)
