--- lexer/test6.c
@@ -1,1 +1,0 @@
-int (*operation)(int x, int y);
