--- lexer/test6_result
@@ -1,13 +1,0 @@
-LexToken(INT,'int',1,0)
-LexToken(LPAREN,'(',1,4)
-LexToken(TIMES,'*',1,5)
-LexToken(ID,'operation',1,6)
-LexToken(RPAREN,')',1,15)
-LexToken(LPAREN,'(',1,16)
-LexToken(INT,'int',1,17)
-LexToken(ID,'x',1,21)
-LexToken(COMMA,',',1,22)
-LexToken(INT,'int',1,24)
-LexToken(ID,'y',1,28)
-LexToken(RPAREN,')',1,29)
-LexToken(SEMI,';',1,30)
