--- lexer/test7.c
@@ -1,1 +1,0 @@
-int v = -3;
